pyproject.toml
[tool.poetry]
name = "found"
version = "0.1.0"
description = ""
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
django-filter = "^25.1"
django-cors-headers = "^4.4"
django-environ = "^0.11"
psycopg2-binary = "^2.9"
gunicorn = "^23.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
